make core tests offline
