+++ sdk/core/azure-core/tests/async_tests/test_pipeline_async.py
@@ -88,54 +88,6 @@ async def test_sans_io_exception():
     with pytest.raises(NotImplementedError):
         await pipeline.run(req)
 
-
-@pytest.mark.asyncio
-async def test_basic_aiohttp():
-
-    request = HttpRequest("GET", "https://bing.com")
-    policies = [
-        UserAgentPolicy("myusergant"),
-        AsyncRedirectPolicy()
-    ]
-    async with AsyncPipeline(AioHttpTransport(), policies=policies) as pipeline:
-        response = await pipeline.run(request)
-
-    assert pipeline._transport.session is None
-    # all we need to check is if we are able to make the call
-    assert isinstance(response.http_response.status_code, int)
-
-@pytest.mark.asyncio
-async def test_basic_aiohttp_separate_session():
-
-    session = aiohttp.ClientSession()
-    request = HttpRequest("GET", "https://bing.com")
-    policies = [
-        UserAgentPolicy("myusergant"),
-        AsyncRedirectPolicy()
-    ]
-    transport = AioHttpTransport(session=session, session_owner=False)
-    async with AsyncPipeline(transport, policies=policies) as pipeline:
-        response = await pipeline.run(request)
-
-    assert transport.session
-    assert isinstance(response.http_response.status_code, int)
-    await transport.close()
-    assert transport.session
-    await transport.session.close()
-
-@pytest.mark.asyncio
-async def test_basic_async_requests():
-
-    request = HttpRequest("GET", "https://bing.com")
-    policies = [
-        UserAgentPolicy("myusergant"),
-        AsyncRedirectPolicy()
-    ]
-    async with AsyncPipeline(AsyncioRequestsTransport(), policies=policies) as pipeline:
-        response = await pipeline.run(request)
-
-    assert isinstance(response.http_response.status_code, int)
-
 @pytest.mark.asyncio
 async def test_async_transport_sleep():
 
@@ -184,33 +136,6 @@ def test_pass_in_http_logging_policy():
     pipeline = pipeline_client._build_pipeline(config)
     http_logging_policy = pipeline._impl_policies[-1]._policy
     assert http_logging_policy.allowed_header_names == HttpLoggingPolicy.DEFAULT_HEADERS_WHITELIST.union({"x-ms-added-header"})
-
-@pytest.mark.asyncio
-async def test_conf_async_requests():
-
-    request = HttpRequest("GET", "https://bing.com/")
-    policies = [
-        UserAgentPolicy("myusergant"),
-        AsyncRedirectPolicy()
-    ]
-    async with AsyncPipeline(AsyncioRequestsTransport(), policies=policies) as pipeline:
-        response = await pipeline.run(request)
-
-    assert isinstance(response.http_response.status_code, int)
-
-def test_conf_async_trio_requests():
-
-    async def do():
-        request = HttpRequest("GET", "https://bing.com/")
-        policies = [
-            UserAgentPolicy("myusergant"),
-            AsyncRedirectPolicy()
-        ]
-        async with AsyncPipeline(TrioRequestsTransport(), policies=policies) as pipeline:
-            return await pipeline.run(request)
-
-    response = trio.run(do)
-    assert isinstance(response.http_response.status_code, int)
 
 @pytest.mark.asyncio
 async def test_retry_without_http_response():
