+++ sdk/core/azure-core/tests/test_custom_hook_policy.py
@@ -3,9 +3,13 @@
 # Licensed under the MIT License.
 # ------------------------------------
 """Tests for the custom hook policy."""
+try:
+    from unittest import mock
+except ImportError:
+    import mock
 from azure.core import PipelineClient
 from azure.core.pipeline.policies import CustomHookPolicy, UserAgentPolicy
-from azure.core.pipeline.transport import HttpRequest
+from azure.core.pipeline.transport import HttpRequest, HttpTransport
 from azure.core.pipeline import PipelineRequest, PipelineContext
 import pytest
 
@@ -13,13 +17,14 @@ def test_response_hook_policy_in_init():
     def test_callback(response):
         raise ValueError()
 
-    url = "https://bing.com"
+    transport = mock.MagicMock(spec=HttpTransport)
+    url = "http://127.0.0.1"
     custom_hook_policy = CustomHookPolicy(raw_response_hook=test_callback)
     policies = [
         UserAgentPolicy("myuseragent"),
         custom_hook_policy
     ]
-    client = PipelineClient(base_url=url, policies=policies)
+    client = PipelineClient(base_url=url, policies=policies, transport=transport)
     request = client.get(url)
     with pytest.raises(ValueError):
         client._pipeline.run(request)
@@ -28,13 +33,14 @@ def test_response_hook_policy_in_request():
     def test_callback(response):
         raise ValueError()
 
-    url = "https://bing.com"
+    transport = mock.MagicMock(spec=HttpTransport)
+    url = "http://127.0.0.1"
     custom_hook_policy = CustomHookPolicy()
     policies = [
         UserAgentPolicy("myuseragent"),
         custom_hook_policy
     ]
-    client = PipelineClient(base_url=url, policies=policies)
+    client = PipelineClient(base_url=url, policies=policies, transport=transport)
     request = client.get(url)
     with pytest.raises(ValueError):
         client._pipeline.run(request, raw_response_hook=test_callback)
@@ -46,13 +52,14 @@ def test_response_hook_policy_in_both():
     def test_callback_request(response):
         raise TypeError()
 
-    url = "https://bing.com"
+    transport = mock.MagicMock(spec=HttpTransport)
+    url = "http://127.0.0.1"
     custom_hook_policy = CustomHookPolicy(raw_response_hook=test_callback)
     policies = [
         UserAgentPolicy("myuseragent"),
         custom_hook_policy
     ]
-    client = PipelineClient(base_url=url, policies=policies)
+    client = PipelineClient(base_url=url, policies=policies, transport=transport)
     request = client.get(url)
     with pytest.raises(TypeError):
         client._pipeline.run(request, raw_response_hook=test_callback_request)
@@ -61,13 +68,14 @@ def test_request_hook_policy_in_init():
     def test_callback(response):
         raise ValueError()
 
-    url = "https://bing.com"
+    transport = mock.MagicMock(spec=HttpTransport)
+    url = "http://127.0.0.1"
     custom_hook_policy = CustomHookPolicy(raw_request_hook=test_callback)
     policies = [
         UserAgentPolicy("myuseragent"),
         custom_hook_policy
     ]
-    client = PipelineClient(base_url=url, policies=policies)
+    client = PipelineClient(base_url=url, policies=policies, transport=transport)
     request = client.get(url)
     with pytest.raises(ValueError):
         client._pipeline.run(request)
@@ -76,13 +84,14 @@ def test_request_hook_policy_in_request():
     def test_callback(response):
         raise ValueError()
 
-    url = "https://bing.com"
+    transport = mock.MagicMock(spec=HttpTransport)
+    url = "http://127.0.0.1"
     custom_hook_policy = CustomHookPolicy()
     policies = [
         UserAgentPolicy("myuseragent"),
         custom_hook_policy
     ]
-    client = PipelineClient(base_url=url, policies=policies)
+    client = PipelineClient(base_url=url, policies=policies, transport=transport)
     request = client.get(url)
     with pytest.raises(ValueError):
         client._pipeline.run(request, raw_request_hook=test_callback)
@@ -94,13 +103,14 @@ def test_request_hook_policy_in_both():
     def test_callback_request(response):
         raise TypeError()
 
-    url = "https://bing.com"
+    transport = mock.MagicMock(spec=HttpTransport)
+    url = "http://127.0.0.1"
     custom_hook_policy = CustomHookPolicy(raw_request_hook=test_callback)
     policies = [
         UserAgentPolicy("myuseragent"),
         custom_hook_policy
     ]
-    client = PipelineClient(base_url=url, policies=policies)
+    client = PipelineClient(base_url=url, policies=policies, transport=transport)
     request = client.get(url)
     with pytest.raises(TypeError):
         client._pipeline.run(request, raw_request_hook=test_callback_request)
