+++ sdk/core/azure-core/tests/test_exceptions.py
@@ -159,17 +159,6 @@ class TestExceptions(object):
         assert isinstance(error.model, FakeErrorTwo)
         assert isinstance(error.error, ODataV4Format)
 
-    def test_httpresponse_error_with_response(self):
-        response = requests.get("https://bing.com")
-        http_response = RequestsTransportResponse(None, response)
-
-        error = HttpResponseError(response=http_response)
-        assert error.message == "Operation returned an invalid status 'OK'"
-        assert error.response is not None
-        assert error.reason == 'OK'
-        assert isinstance(error.status_code, int)
-        assert error.error is None
-
     def test_odata_v4_exception(self):
         message = {
             "error": {
