+++ sdk/core/azure-core/tests/testserver_tests/async_tests/test_pipeline_async_test_server.py
@@ -1,0 +1,132 @@
+#--------------------------------------------------------------------------
+#
+# Copyright (c) Microsoft Corporation. All rights reserved.
+#
+# The MIT License (MIT)
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the ""Software""), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+#
+#--------------------------------------------------------------------------
+import sys
+
+from azure.core.pipeline import AsyncPipeline
+from azure.core.pipeline.policies import (
+    SansIOHTTPPolicy,
+    UserAgentPolicy,
+    DistributedTracingPolicy,
+    AsyncRetryPolicy,
+    AsyncRedirectPolicy,
+    AsyncHTTPPolicy,
+    AsyncRetryPolicy,
+    HttpLoggingPolicy,
+)
+from azure.core.pipeline.transport import (
+    AsyncHttpTransport,
+    HttpRequest,
+    AsyncioRequestsTransport,
+    TrioRequestsTransport,
+    AioHttpTransport
+)
+
+from azure.core.polling.async_base_polling import AsyncLROBasePolling
+from azure.core.polling.base_polling import LocationPolling
+
+from azure.core.configuration import Configuration
+from azure.core import AsyncPipelineClient
+from azure.core.exceptions import AzureError
+
+import aiohttp
+import trio
+
+import pytest
+
+
+@pytest.mark.asyncio
+async def test_basic_aiohttp(port):
+
+    request = HttpRequest("GET", "http://localhost:{}/basic/string".format(port))
+    policies = [
+        UserAgentPolicy("myusergant"),
+        AsyncRedirectPolicy()
+    ]
+    async with AsyncPipeline(AioHttpTransport(), policies=policies) as pipeline:
+        response = await pipeline.run(request)
+
+    assert pipeline._transport.session is None
+    # all we need to check is if we are able to make the call
+    assert isinstance(response.http_response.status_code, int)
+
+@pytest.mark.asyncio
+async def test_basic_aiohttp_separate_session(port):
+
+    session = aiohttp.ClientSession()
+    request = HttpRequest("GET", "http://localhost:{}/basic/string".format(port))
+    policies = [
+        UserAgentPolicy("myusergant"),
+        AsyncRedirectPolicy()
+    ]
+    transport = AioHttpTransport(session=session, session_owner=False)
+    async with AsyncPipeline(transport, policies=policies) as pipeline:
+        response = await pipeline.run(request)
+
+    assert transport.session
+    assert isinstance(response.http_response.status_code, int)
+    await transport.close()
+    assert transport.session
+    await transport.session.close()
+
+@pytest.mark.asyncio
+async def test_basic_async_requests(port):
+
+    request = HttpRequest("GET", "http://localhost:{}/basic/string".format(port))
+    policies = [
+        UserAgentPolicy("myusergant"),
+        AsyncRedirectPolicy()
+    ]
+    async with AsyncPipeline(AsyncioRequestsTransport(), policies=policies) as pipeline:
+        response = await pipeline.run(request)
+
+    assert isinstance(response.http_response.status_code, int)
+
+@pytest.mark.asyncio
+async def test_conf_async_requests(port):
+
+    request = HttpRequest("GET", "http://localhost:{}/basic/string".format(port))
+    policies = [
+        UserAgentPolicy("myusergant"),
+        AsyncRedirectPolicy()
+    ]
+    async with AsyncPipeline(AsyncioRequestsTransport(), policies=policies) as pipeline:
+        response = await pipeline.run(request)
+
+    assert isinstance(response.http_response.status_code, int)
+
+def test_conf_async_trio_requests(port):
+
+    async def do():
+        request = HttpRequest("GET", "http://localhost:{}/basic/string".format(port))
+        policies = [
+            UserAgentPolicy("myusergant"),
+            AsyncRedirectPolicy()
+        ]
+        async with AsyncPipeline(TrioRequestsTransport(), policies=policies) as pipeline:
+            return await pipeline.run(request)
+
+    response = trio.run(do)
+    assert isinstance(response.http_response.status_code, int)
