+++ sdk/core/azure-core/tests/testserver_tests/test_exceptions_test_server.py
@@ -1,0 +1,40 @@
+# --------------------------------------------------------------------------
+#
+# Copyright (c) Microsoft Corporation. All rights reserved.
+#
+# The MIT License (MIT)
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the ""Software""), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+#
+# --------------------------------------------------------------------------
+import requests
+from azure.core.exceptions import HttpResponseError
+from azure.core.pipeline.transport import RequestsTransportResponse
+
+class TestExceptions(object):
+    def test_httpresponse_error_with_response(self, port):
+        response = requests.get("http://localhost:{}/basic/string".format(port))
+        http_response = RequestsTransportResponse(None, response)
+
+        error = HttpResponseError(response=http_response)
+        assert error.message == "Operation returned an invalid status 'OK'"
+        assert error.response is not None
+        assert error.reason == 'OK'
+        assert isinstance(error.status_code, int)
+        assert error.error is None
